Locally fork AOSP master vulkan headers
We can't touch the platform's copy of the headers because they're used
to generate libvulkan.so in the system and vendor images, and this is
NDK ABI stable. Since gfxstream needs just the headers and happily uses
them from another C path (this is done for the emulator anyway), just
fork the headers from AOSP master into the android11-gsi version.

Bug: 216827224
Change-Id: Ia3e0864b0a6e18e1d8be7d2d14e1256dc1f43d2f