goldfish-opengl: threadID --> resourceId
The virtgpu resource ID is guranteed to be an system wide unique
identifier.  Since all ASG ring buffers are backed by host coherent
memory, it makes to identify the ring based on the resource ID.
We can also use gettid(..) as well, but since we need to tie the
lifetime of the ASG external blob to the virtgpu resource, resource
ID along does the trick.

Bug: b/297561291
Test: launch_cvd --gpu_mode=gfxstream

BUG=235485545
TEST=compile
Change-Id: I1b3c417358abd499186ba374935d79756648602b

Change-Id: Ia29866d7168f2de21c1753aa979f249c953de488