hw3:  add ro.root.qemu.external.displays
to the list of
hwservice.external.displays; now emulator can do the
following so the external displays can be added without
rebuilding the system image. for example,
adding display 1 with 1080x1920 and 420 density with
default flag.
emulator \
-append-userspace-opt \
androidboot.qemu.external.displays="1,1080,2092,420,0"

Test:
launch emulator with the above option will produce
two Built-in displays with the second one as 1080x2092x420

Bug: 289373564
Change-Id: I8a37cfae9ad8a00e87fed29ee720399171d081c6