DrmClient: check card0 then card1 for virtio-gpu
This modification is intended to work around a problem where the
kernel might load the vkms module and initialize it as card0.

If for whatever reason card0 is not virtio-gpu, then we attempt to
open card1 instead of returning an error right away. If card1 is
also not virtio-gpu, we return an error.

Change-Id: Ib6167baa04468c734d1402d3dc07ea4b39c7dd7d