codecs: try to fix android.mediav2.cts.DecoderDynamicColorAspectTest
Note: vp8 and vp9 are similar, but there are important differences

vp9 has FLAG_CODEC_CONFIG input frame, and when this frame comes,
we can assume there is config changes, mostly in this test,
there will be color space configuraiton changes and we can take
this moment to propogate the default config to the real color
space. Since vp9 frame itself does not carry coded color space
config, we just copy default to be the one used.

However vp8 does not have FLAG_CODEC_CONFIG frame, and this makes
it hard to tell when there is a change of color space configuration.
For this specific test, we simply check the time stamp changes
compared to previous one, and when it is more than 1 second, we
will re-populate default color space to the real color space.

In addition, we need to create this color space and stick it to
each decoded output buffer for vp8, to avoid race condition when
setting and accessing the real color space data, as it is a
shared pointer and can change across different thread.

Ignore-AOSP-First: ct burn down

Bug: 349159609

Change-Id: I614b7659db8bf01af5ae912d2c8e10c5a163c5a2